+++ crates/rune-macros/src/context.rs
@@ -625,11 +625,11 @@ impl Context {
             full_type_of: quote!(#module::runtime::FullTypeOf),
             hash: quote!(#module::Hash),
             id: quote!(#module::parse::Id),
-            install_with: quote!(#module::module::InstallWith),
+            install_with: quote!(#module::__private::InstallWith),
             macro_context: quote!(#module::macros::MacroContext),
             maybe_type_of: quote!(#module::runtime::MaybeTypeOf),
             module_variant: quote!(#module::module::Variant),
-            module: quote!(#module::module::Module),
+            module: quote!(#module::__private::Module),
             named: quote!(#module::compile::Named),
             object: quote!(#module::runtime::Object),
             opaque: quote!(#module::parse::Opaque),
